StatsService with 6 methods and tests

+++ src/main/java/ru/netology/stats/StatsService.java
@@ -2,21 +2,22 @@ package ru.netology.stats;
 
 public class StatsService {
 
-    public int sumSales(int[] sales) {
-        int sum = 0;
-        for (int sale : sales) {
+
+    public long sumSales(long[] sales) {
+        long sum = 0;
+        for (long sale : sales) {
             sum += sale;
         }
         return sum;
     }
 
 
-    public int averageSales(int[] sales) {
+    public long averageSales(long[] sales) {
         return sumSales(sales) / sales.length;
     }
 
 
-    public int maxSalesMonth(int[] sales) {
+    public int maxSalesMonth(long[] sales) {
         int maxMonth = 0;
         for (int i = 0; i < sales.length; i++) {
             if (sales[i] >= sales[maxMonth]) {
@@ -27,7 +28,7 @@ public class StatsService {
     }
 
 
-    public int minSalesMonth(int[] sales) {
+    public int minSalesMonth(long[] sales) {
         int minMonth = 0;
         for (int i = 0; i < sales.length; i++) {
             if (sales[i] <= sales[minMonth]) {
@@ -38,10 +39,10 @@ public class StatsService {
     }
 
 
-    public int monthsBelowAverage(int[] sales) {
-        int average = averageSales(sales);
+    public int monthsBelowAverage(long[] sales) {
+        long average = averageSales(sales);
         int count = 0;
-        for (int sale : sales) {
+        for (long sale : sales) {
             if (sale < average) {
                 count++;
             }
@@ -50,10 +51,10 @@ public class StatsService {
     }
 
 
-    public int monthsAboveAverage(int[] sales) {
-        int average = averageSales(sales);
+    public int monthsAboveAverage(long[] sales) {
+        long average = averageSales(sales);
         int count = 0;
-        for (int sale : sales) {
+        for (long sale : sales) {
             if (sale > average) {
                 count++;
             }
