+++ src/test/java/ru/netology/stats/StatsServiceTest.java
@@ -1,26 +1,25 @@
 package ru.netology.stats;
 
 import org.junit.jupiter.api.Test;
-
 import static org.junit.jupiter.api.Assertions.*;
 
 public class StatsServiceTest {
 
-    int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+    long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
 
     @Test
     void sumSales() {
         StatsService service = new StatsService();
-        int expected = 180;
-        int actual = service.sumSales(sales);
+        long expected = 180;
+        long actual = service.sumSales(sales);
         assertEquals(expected, actual);
     }
 
     @Test
     void averageSales() {
         StatsService service = new StatsService();
-        int expected = 15;
-        int actual = service.averageSales(sales);
+        long expected = 15;
+        long actual = service.averageSales(sales);
         assertEquals(expected, actual);
     }
